Handle the Properties and Models of various boards
The detection of information of the card in boards.txt failed when
the information were assigned to families and specific models

We used to look only at the properties of a specific model but sometimes
we also have to check for properties in the "parent" node.

This patch corrects that by allowing to specify both the family and
the model of the card if required.

Example of what the patch corrects:

        mega.build.f_cpu=16000000L
        mega.menu.cpu.atmega2560.build.mcu=atmega2560

For this card, the F_CPU is given by mega (the family) and the MCU is
given by the specific model (atmega2560).